django/requirements.txt
awsebcli==3.7.7
blosc==1.3.3
botocore==1.4.33
decorator==3.4.0
Django==1.9.7
django-rest-swagger==2.0.1
djangorestframework==3.4
docutils==0.12
numpy==1.11.1
pathspec==0.3.4
psycopg2==2.6.2
PyYAML==3.11
requests==2.9.1
six==1.10.0
virtualenv==15.0.2
